Docs, packaging fixes

+++ run_docs.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-pdoc --force --html --output-dir ./docs qose
-cd ./docs/qose
-mv * ../
-cd ..
-rm -r ./qose/
+if pdoc --force --html --output-dir ./docs ./qose --skip-errors; then
+    cd ./docs/qose
+    mv * ../
+    cd ..
+    rm -r ./qose/
+    echo "Docs generated!"
+else
+    echo "Docs failed!"
+fi
+
